Replace timer with after-change-functions
Timer is runs all the time, even when there are no changes. This means that
when the point is on the base buffer, the org-tree window is constantly
refreshing. This makes scrolling the unselected org-tree buffer not work

This problem also occurs with post-command-hook.

Fixed by using after-change-functions.

However, cursor moves don't count as changes, so the line doesn't move
to a new heading when the cursor moves in the base buffer.

Maybe that's ok?