Use post-command-hook instead of after-change-hook
Org-metadown caused a lot of problems, since it enacted a lot of changes,
thus triggering several intermediate updates. It would always move point back
to the right heading, but cursor moves don't trigger the
after-change-functions hook.

Post-command-hook solves this problem, although it has the same issue as the
timer; scrolling the tree buffer with point still in base buffer triggered
updates, and therefore disallowed scrolling.

Added some logic to the hook function to track the last cursor point with a
buffer-local var, and only update when cursor has moved.